% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GOCore.R
\name{bezier}
\alias{bezier}
\title{bezier}
\usage{
bezier(data, process.col)
}
\arguments{
\item{data}{data frame x.start,x.end,y.start,y.end}

\item{process.col}{Color of the processes}
}
\description{
bezier
}

