% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GOCore.R
\name{circle_dat}
\alias{circle_dat}
\title{Creates a plotting object.}
\usage{
circle_dat(terms, genes)
}
\arguments{
\item{terms}{A data frame with columns for 'category', 'ID', 'term', adjusted
p-value ('adj_pval') and 'genes'}

\item{genes}{A data frame with columns for 'ID', 'logFC'}
}
\description{
The function takes the results from a functional analysis (for
  example DAVID) and combines it with a list of selected genes and their
  logFC. The resulting data frame can be used as an input for various ploting
  functions.
}
\details{


The circ object has eight columns with the following names:
    * category
    * ID
}
\examples{
\dontrun{
#Load the included dataset
data(EC)

#Building the circ object
circ<-circular_dat(EC$david, EC$genelist)
}
}

