% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOCore.R
\docType{data}
\name{EC}
\alias{EC}
\title{Transcriptomic information of endothelial cells.}
\format{A list containing 5 items}
\source{
\url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE47067}
}
\usage{
data(EC)
}
\description{
The data set contains the transcriptomic information of endothelial cells
from two steady state tissues (brain and heart). More detailed information
can be found in the paper by Nolan et al. 2013. The data was normalized and a
statistical analysis was performed to determine differentially expressed
genes. DAVID functional annotation tool was used to perform a gene-
annotation enrichment analysis of the set of differentially expressed genes
(adjusted p-value < 0.05).
}
\keyword{datasets}

