% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GOVenn.R
\name{get_overlap}
\alias{get_overlap}
\title{get_overlap}
\usage{
get_overlap(A, B, C)
}
\arguments{
\item{A}{Input list 1}

\item{B}{Input list 2}

\item{C}{Input list 3}
}
\description{
get_overlap
}

